Rename 'jdbc' storage to 'jpa'
