+++ app/src/main/java/io/apicurio/registry/storage/impl/jpa/JPAEntityMapper.java
@@ -1,12 +1,12 @@
-package io.apicurio.registry.storage.impl.jdbc;
+package io.apicurio.registry.storage.impl.jpa;
 
 import io.apicurio.registry.storage.StoredArtifact;
-import io.apicurio.registry.storage.impl.jdbc.entity.Artifact;
+import io.apicurio.registry.storage.impl.jpa.entity.Artifact;
 
 import javax.enterprise.context.Dependent;
 
 @Dependent
-public class JdbcEntityMapper {
+public class JPAEntityMapper {
 
     public StoredArtifact toStoredArtifact(Artifact artifact) {
         return StoredArtifact.builder()
